Struct math::big::Int

pub struct Int(_);

An Int represents a signed multi-precision integer. The zero value for an Int represents the value 0.

Operations always take mutable reference arguments (&mut Int) rather than Int values, and each unique Int value requires its own unique &Int reference. To “copy” an Int value, just use Int::clone.

Implementations

impl Int

pub fn abs(&mut self, x: &Self) -> &mut Self

abs sets self to |x| (the absolute value of x) and returns self.

pub fn add(&mut self, x: &Self, y: &Self) -> &mut Self

Sets self to the sum x+y and returns self.

pub fn and(&mut self, x: &Self, y: &Self) -> &mut Self

Sets self = x & y and returns self.

pub fn and_not(&mut self, x: &Self, y: &Self) -> &mut Self

Sets self = x &^ y and returns self.

pub fn append(&mut self, buf: Vec<u8>, base: u32) -> Vec<u8>

Appends the string representation of self, as generated by self.text(base), to buf and returns the extended buffer.

pub fn binomial(&mut self, n: i64, k: i64) -> &mut Self

Sets self to the binomial coefficient of (n, k) and returns self.

pub fn bit(&self, i: usize) -> u8

Returns the value of the i’th bit of self. That is, it returns (self>>i)&1. The bit index i must be >= 0.

pub fn bit_len(&self) -> usize

Returns the length of the absolute value of self in bits.

pub fn bits(&self) -> Box<dyn DoubleEndedIterator<Item = u32> + '_>

Provides raw (unchecked but fast) access to self by returning its absolute value as a little-endian Word slice. The result and self share the same underlying array. Bits is intended to support implementation of missing low-level Int functionality outside this package; it should be avoided otherwise.

Use the iterator in the reversed direction will produce a big-endian Word slice.

pub fn bytes(&self) -> Vec<u8>

Returns the absolute value of self as a big-endian byte slice.

To use a fixed length slice, or a preallocated one, use Int::fill_bytes.

pub fn cmp(&self, y: &Self) -> i32

Compares x and y and returns:

-1 if x <  y
 0 if x == y
+1 if x >  y

pub fn cmp_abs(&self, y: &Self) -> i32

Compares the absolute values of x and y and returns:

-1 if |x| <  |y|
 0 if |x| == |y|
+1 if |x| >  |y|

pub fn div(&mut self, x: &Self, y: &Self) -> &mut Self

Sets self to the quotient x/y for y != 0 and returns self. If y == 0, a division-by-zero run-time panic occurs. div implements Euclidean division (unlike Go); see Int::div_mod for more details.

pub fn div_mod<'a, 'b>(
&'a mut self,
x: &Self,
y: &Self,
m: &'b mut Self
) -> (&'a mut Self, &'b mut Self)

Sets self to the quotient x div y and m to the modulus x mod y and returns the pair (self, m) for y != 0. If y == 0, a division-by-zero run-time panic occurs.

div_mod implements Euclidean division and modulus (unlike Go):

q = x div y  such that
m = x - y*q  with 0 <= m < |y|

(See Raymond T. Boute, “The Euclidean definition of the functions div and mod”. ACM Transactions on Programming Languages and Systems (TOPLAS), 14(2):127-144, New York, NY, USA, 4/1992. ACM press.) See Int::quo_rem for T-division and modulus (like Go).

pub fn exp(&mut self, x: &Self, y: &Self, m: Option<&Self>) -> Option<&mut Self>

Sets self = x**y mod |m| (i.e. the sign of m is ignored), and returns self. If m == nil or m == 0, self = x**y unless y <= 0 then self = 1. If m != 0, y < 0, and x and m are not relatively prime, self is unchanged and None is returned.

Modular exponentiation of inputs of a particular size is not a cryptographically constant-time operation.

pub fn fill_bytes<'a>(&self, out: &'a mut [u8]) -> &'a mut [u8]

Sets buf to the absolute value of self, storing it as a zero-extended big-endian byte slice, and returns buf.

If the absolute value of self doesn’t fit in buf, Int::fill_bytes will panic.

pub fn gcd(
&mut self,
x: Option<&mut Self>,
y: Option<&mut Self>,
a: &Self,
b: &Self
) -> &mut Self

Sets self to the greatest common divisor of a and b and returns z. If x or y are not None, GCD sets their value such that z = a*x + b*y.

a and b may be positive, zero or negative. (Before Go 1.14 both had to be > 0.) Regardless of the signs of a and b, self is always >= 0.

If a == b == 0, GCD sets z = x = y = 0.

If a == 0 and b != 0, GCD sets z = |b|, x = 0, y = sign(b) * 1.

If a != 0 and b == 0, GCD sets z = |a|, x = sign(a) * 1, y = 0.

pub fn int64(&self) -> i64

Returns the i64 representation of self. If self cannot be represented in an i64, the result is undefined.

pub fn is_int64(&self) -> bool

Reports whether self can be represented as an int64.

pub fn is_uint64(&self) -> bool

Reports whether self can be represented as a uint64.

pub fn lsh(&mut self, x: &Self, n: usize) -> &mut Self

Sets self = x << n and returns self.

pub fn mod(&mut self, x: &Self, y: &Self) -> &mut Self

Sets self to the modulus x%y for y != 0 and returns self. If y == 0, a division-by-zero run-time panic occurs. mod implements Euclidean modulus (unlike Go); see Int::div_mod for more details.

pub fn mod_inverse(&mut self, g: &Self, n: &Self) -> Option<&mut Self>

Sets self to the multiplicative inverse of g in the ring ℤ/nℤ and returns self. If g and n are not relatively prime, g has no multiplicative inverse in the ring ℤ/nℤ. In this case, self is unchanged and the return value is None. If n == 0, a division-by-zero run-time panic occurs.

pub fn mod_sqrt(&mut self, x: &Self, p: &Self) -> Option<&mut Self>

Sets self to a square root of self mod p if such a square root exists, and returns self. The modulus p must be an odd prime. If x is not a square mod p, mod_sqrt leaves self unchanged and returns nil. This function panics if p is not an odd integer, its behavior is undefined if p is odd but not prime.

pub fn mul(&mut self, x: &Self, y: &Self) -> &mut Self

Sets self to the product x*y and returns self.

pub fn mul_range(&mut self, a: i64, b: i64) -> &mut Self

Sets self to the product of all integers in the range [a, b] inclusively and returns self. If a > b (empty range), the result is 1.

pub fn neg(&mut self, x: &Self) -> &mut Self

Sets self to -x and returns self.

pub fn not(&mut self, x: &Self) -> &mut Self

Sets z = ^x and returns self.

pub fn or(&mut self, x: &Self, y: &Self) -> &mut Self

Sets self = x | y and returns self.

pub fn quo(&mut self, x: &Self, y: &Self) -> &mut Self

Sets self to the quotient x/y for y != 0 and returns self. If y == 0, a division-by-zero run-time panic occurs. quo implements truncated division (like Go); see Int::quo_rem for more details.

pub fn quo_rem<'a, 'b>(
&'a mut self,
x: &Self,
y: &Self,
r: &'b mut Self
) -> (&'a mut Self, &'b mut Self)

Sets self to the quotient x/y and r to the remainder x%y and returns the pair (self, r) for y != 0. If y == 0, a division-by-zero run-time panic occurs.

quo_rem implements T-division and modulus (like Go):

q = x/y      with the result truncated to zero
r = x - y*q

(See Daan Leijen, “Division and Modulus for Computer Scientists”.) See DivMod for Euclidean division and modulus (unlike Go).

pub fn rand<R>(&mut self, rnd: &mut R, n: &Self) -> &mut Selfwhere
R: Read,

Sets self to a pseudo-random number in [0, n) and returns self.

Security depends on the randomness provided by rnd.

pub fn rem(&mut self, x: &Self, y: &Self) -> &mut Self

Sets self to the remainder x%y for y != 0 and returns self. If self == 0, a division-by-zero run-time panic occurs. rem implements truncated modulus (like Go); see Int::quo_rem for more details.

pub fn rsh(&mut self, x: &Self, n: usize) -> &mut Self

Sets self = x >> n and returns self.

pub fn set(&mut self, x: &Self) -> &mut Self

Sets self to x and returns self.

pub fn set_bit(&mut self, x: &Self, i: usize, b: bool) -> &mut Self

Sets self to x, with x’s i’th bit set to 0/1 for b=false/true. That is, if b is true set_bit sets self = x | (1 << i); if b is false set_bit sets self = x &^ (1 << i).

pub fn set_bits(&mut self, abs: &[u32]) -> &mut Self

Provides raw (unchecked but fast) access to self by setting its value to abs, interpreted as a little-endian Word slice, and returning self. set_bits is intended to support implementation of missing low-level Int functionality outside this package; it should be avoided otherwise.

pub fn set_bytes(&mut self, buf: &[u8]) -> &mut Self

Interprets buf as the bytes of a big-endian unsigned integer, sets self to that value, and returns self.

pub fn set_int64(&mut self, x: i64) -> &mut Self

Sets self to x and returns self.

pub fn set_string(&mut self, s: &str, base: u8) -> Option<&mut Self>

Sets self to the value of s, interpreted in the given base, and returns self and Option<&mut Self> indicating success. The entire string (not just a prefix) must be valid for success. If set_string fails, the value of self is undefined but the returned value is None.

The base argument must be 0 or a value between 2 and MAX_BASE. For base 0, the number prefix determines the actual base: A prefix of “0b” or “0B” selects base 2, “0”, “0o” or “0O” selects base 8, and “0x” or “0X” selects base 16. Otherwise, the selected base is 10 and no prefix is accepted.

For bases <= 36, lower and upper case letters are considered the same: The letters ‘a’ to ‘z’ and ‘A’ to ‘Z’ represent digit values 10 to 35. For bases > 36, the upper case letters ‘A’ to ‘Z’ represent the digit values 36 to 61.

For base 0, an underscore character “_” may appear between a base prefix and an adjacent digit, and between successive digits; such underscores do not change the value of the number. Incorrect placement of underscores is reported as an error if there are no other errors. If base != 0, underscores are not recognized and act like any other character that is not a valid digit.

Example

use math::big::Int;

fn main() {
    let mut i = Int::default();
    i.set_string("644", 8);

    assert_eq!(i, Int::new(420));
}

pub fn set_uint64(&mut self, x: u64) -> &mut Self

Sets self to x and returns self.

pub fn sign(&self) -> i32

Returns:

-1 if x <  0
 0 if x == 0
+1 if x >  0

pub fn sqrt(&mut self, x: &Self) -> &mut Self

Sets self to ⌊√x⌋, the largest integer such that self² ≤ x, and returns self. It panics if self is negative.

pub fn string(&self) -> String

Returns the decimal representation of self as generated by self.text(10). This method is provided just for consistent with golang. Use self.to_string instead.

pub fn sub(&mut self, x: &Self, y: &Self) -> &mut Self

Sets self to the difference x-y and returns self.

pub fn text(&self, base: u32) -> String

Returns the string representation of self in the given base. Base must be between 2 and 62, inclusive. The result uses the lower-case letters ‘a’ to ‘z’ for digit values 10 to 35, and the upper-case letters ‘A’ to ‘Z’ for digit values 36 to 61. No prefix (such as “0x”) is added to the string.

pub fn trailing_zero_bits(&self) -> usize

Returns the number of consecutive least significant zero bits of |self|.

pub fn uint64(&self) -> u64

returns the uint64 representation of self. If self cannot be represented in a uint64, the result is undefined.

pub fn xor(&mut self, x: &Self, y: &Self) -> &mut Self

Sets self = x ^ y and returns self.

pub fn new(x: i64) -> Self

Allocates and returns a new Int set to x.