Module math::bits

Constants

UINT_SIZE

Functions

add

add returns the sum with carry of x, y and carry: sum = x + y + carry. The carry input must be 0 or 1; otherwise the behavior is undefined. The carryOut output is guaranteed to be 0 or 1.

add32

add32 returns the sum with carry of x, y and carry: sum = x + y + carry. The carry input must be 0 or 1; otherwise the behavior is undefined. The carryOut output is guaranteed to be 0 or 1.

add64

add64 returns the sum with carry of x, y and carry: sum = x + y + carry. The carry input must be 0 or 1; otherwise the behavior is undefined. The carryOut output is guaranteed to be 0 or 1.

div

div returns the quotient and remainder of (hi, lo) divided by y: quo = (hi, lo)/y, rem = (hi, lo)%y with the dividend bits’ upper half in parameter hi and the lower half in parameter lo. div panics for y == 0 (division by zero) or y <= hi (quotient overflow).

div32

div32 returns the quotient and remainder of (hi, lo) divided by y: quo = (hi, lo)/y, rem = (hi, lo)%y with the dividend bits’ upper half in parameter hi and the lower half in parameter lo. div32 panics for y == 0 (division by zero) or y <= hi (quotient overflow).

div64

div64 returns the quotient and remainder of (hi, lo) divided by y: quo = (hi, lo)/y, rem = (hi, lo)%y with the dividend bits’ upper half in parameter hi and the lower half in parameter lo. div64 panics for y == 0 (division by zero) or y <= hi (quotient overflow).

sub

sub returns the difference of x, y and borrow: diff = x - y - borrow. The borrow input must be 0 or 1; otherwise the behavior is undefined. The borrowOut output is guaranteed to be 0 or 1.

sub32

sub64