math::kinematics
Trait Kinematics

pub trait Kinematics {
    type State;

    // Required methods
    fn inverse(&self, robot_speeds: Vector3<f32>) -> Self::State;
    fn forward(&self, state: Self::State) -> Vector3<f32>;
}

A trait that takes in a the requested robot speeds and returns the state the drivetrain should be in to achieve those speeds.

The vector is [x, y, rotation] in the WPILib robot coordinate system (Link to WPILib docs).

Required Associated Types

type State

Required Methods

fn inverse(&self, robot_speeds: Vector3<f32>) -> Self::State

Convert from robot speeds to drive train state

fn forward(&self, state: Self::State) -> Vector3<f32>

Convert from drive train state to robot speeds